requirements.txt
# Data Architecture for AI — Python dependencies
# pip install -r requirements.txt

boto3>=1.34.0
great-expectations>=0.18.0
numpy>=1.26.0
opensearch-py>=2.4.0
gremlin-python>=3.7.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
requests-aws4auth>=1.3.0
